Fix tasks

+++ CryptoStat/AuthAndStat/models.py
@@ -4,4 +4,4 @@ from django.db import models
 class Info(models.Model):
     date = models.DateField()
     currency = models.CharField(max_length=3)
-    equity = models.IntegerField()
+    equity = models.FloatField()
